adding changes

+++ prestamosAPI/src/main/java/com/prestamo/entities/CuotaPrestamo.java
@@ -4,7 +4,7 @@ public class CuotaPrestamo {
 	private int idPrestamo;
 	private int idCliente;
 	private int numeroCuota;
-	private int monto;
+	private double monto;
 	private String fechaVencimiento;
  	public CuotaPrestamo() {
 		super();
@@ -42,11 +42,11 @@ public class CuotaPrestamo {
 		this.numeroCuota = numeroCuota;
 	}
 
-	public int getMonto() {
+	public double getMonto() {
 		return monto;
 	}
 
-	public void setMonto(int monto) {
+	public void setMonto(double monto) {
 		this.monto = monto;
 	}
